GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - features (refs/gitbutler/features)
   branch head: 6771164dce039325580169a41cbc40082d8d6336
   - components/ui/header.tsx
   - package-lock.json
   - pages/_app.tsx
   - components/CookieBanner.tsx
   - app/servicios/page.tsx
   - components/ui/mobile-menu.tsx
   - components/zigzag.tsx
   - package.json
   - components/hero.tsx
   - app/nosotros/page.tsx
   - public/images/Free whatsapp blue logo vector.svg-svg.svg
   - public/images/IBM_logo.svg.png
   - public/images/SaucotecLogo-white-svg.svg
   - tailwind.config.js
   - app/clientes/page.tsx
   - .vscode/settings.json
   - app/css/style.css
   - public/images/Microsoft_logo_(2012).svg.png
   - public/images/640px-Apple-logo.png

Your previous branch was: refs/heads/features

The sha for that commit was: 6771164dce039325580169a41cbc40082d8d6336

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch