GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - HEAD (refs/gitbutler/head)
   - components/ui/header.tsx
   - app/nosotros/page.tsx
   - app/servicios/page.tsx

Your previous branch was: HEAD

The sha for that commit was: cf8441e4eed31eb02404c232305f13bc326ac4a7

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch